// frontend/src/types/database.types.ts
export interface NoteDB {
  id: string;
  date_entered: Date | string | null;
  date_modified: Date | string | null;
  created_by: string | null;
  modified_user_id: string | null;
  assigned_user_id: string | null;
  deleted: number | null;
  name: string | null;
  parent_type: string | null;
  parent_id: string | null;
  contact_id: string | null;
  description: string | null;
}